Add a readme that explains some of the changes that are on the way
