+++ remix-vendor/cdk/index.ts
@@ -2,5 +2,12 @@ import "./globals";
 
 export { createArcTableSessionStorage } from "./sessions/dynamoTableSessionStorage";
 
+export {
+  capturedClient,
+  documentClient,
+  isLambdaEnvironment,
+} from "./utils/aws";
+
 export type { GetLoadContextFunction, RequestHandler } from "./server";
+
 export { createRequestHandler } from "./server";
