+++ remix-vendor/cdk/sessions/dynamoTableSessionStorage.ts
@@ -1,45 +1,9 @@
-import * as crypto from "crypto";
-import type { SessionStorage, SessionIdStorageStrategy } from "@remix-run/node";
+import type { SessionStorage } from "@remix-run/node";
 import { createSessionStorage } from "@remix-run/node";
 import { SessionData } from "@remix-run/server-runtime";
-import AWS, { DynamoDB } from "aws-sdk";
-import { captureAWSClient } from "aws-xray-sdk";
-import { isLambdaEnvironment } from "../utils/aws";
-
-AWS.config.logger = console;
-
-const baseClient = new DynamoDB();
-
-const capturedClient = isLambdaEnvironment
-  ? captureAWSClient(baseClient)
-  : baseClient;
-
-const documentClient = new DynamoDB.DocumentClient(capturedClient);
-
-interface ArcTableSessionStorageOptions {
-  /**
-   * The name of the DynamoDB table used for storing sessions
-   */
-  tableName: string;
-
-  /**
-   * The name of the DynamoDB attribute used to store the session ID.
-   * This should be the table's partition key.
-   */
-  idx: string;
-
-  /**
-   * The Cookie used to store the session id on the client, or options used
-   * to automatically create one.
-   */
-  cookie?: SessionIdStorageStrategy["cookie"];
-
-  /**
-   * The name of the DynamoDB attribute used to store the expiration time.
-   * If absent, then no TTL will be stored and session records will not expire.
-   */
-  ttl?: string;
-}
+import * as crypto from "crypto";
+import { documentClient } from "../utils/aws";
+import { IArcTableSessionStorageOptions } from "./types";
 
 /**
  * Session storage using a DynamoDB table managed by aws-sdk.
@@ -51,12 +15,11 @@ interface ArcTableSessionStorageOptions {
 export function createArcTableSessionStorage({
   cookie,
   ...props
-}: ArcTableSessionStorageOptions): SessionStorage {
+}: IArcTableSessionStorageOptions): SessionStorage {
   return createSessionStorage({
     cookie,
 
     async createData(data, expires) {
-      console.info("Create data", { props });
       while (true) {
         let randomBytes = crypto.randomBytes(8);
         // This storage manages an id space of 2^64 ids, which is far greater
@@ -101,7 +64,6 @@ export function createArcTableSessionStorage({
     },
 
     async readData(id) {
-      console.info("Read data", { props });
       let { Item } = await documentClient
         .get({
           TableName: props.tableName,
@@ -120,7 +82,6 @@ export function createArcTableSessionStorage({
     },
 
     async updateData(id, data, expires) {
-      console.info("Update data", { props });
       let params = {
         [props.idx]: id,
         ...data,
@@ -139,7 +100,6 @@ export function createArcTableSessionStorage({
     },
 
     async deleteData(id) {
-      console.info("Delete data", { props });
       await documentClient.delete({
         TableName: props.tableName,
         Key: {
